+++ P-A84/pins_config.cpp
@@ -1,14 +1,14 @@
 #include "pins.h"
 
-const int rowDataPin = 5;
-const int rowClockPin = 0;
-const int rowLatchPin = 1;
+const short int rowDataPin = 5;
+const short int rowClockPin = 0;
+const short int rowLatchPin = 1;
 
-const int colDataPin = 6;
-const int colClockPin = 3;
-const int colLatchPin = 2;
+const short int colDataPin = 6;
+const short int colClockPin = 3;
+const short int colLatchPin = 2;
 
-const int btnLeftPin = 10;
-const int btnOkPin = 9;
-const int btnCancelPin = 8;
-const int btnRightPin = 7;
+const short int btnLeftPin = 10;
+const short int btnOkPin = 9;
+const short int btnCancelPin = 8;
+const short int btnRightPin = 7;
